Trivial bits
git-svn-id: svn://svn.lyx.org/lyx/lyx-devel/trunk@5292 a592a061-630c-0410-9148-cb99ea01b6c8

+++ src/frontends/qt2/iconpalette.C
@@ -72,7 +72,7 @@ void IconPalette::resizeEvent(QResizeEvent * e)
 		return;
 
 	int cols(width() / button_size);
-	int rows = max(buttons_.size() / cols, 1U);
+	int rows = max(int(buttons_.size() / cols), 1);
 	if (buttons_.size() % cols)
 		++rows;
